remove mergetContext
